chore: avoid macros with memory clobber
sei() and cli() apparently have "::memory" clobbers. Avoid that in the
context switches.

+++ simplos/src/scheduler.c
@@ -30,7 +30,7 @@ void reschedule(void) {
 
   prioritize_tasks(kernel->schedule.queue.tasks, num_active_tasks,
                    kernel->schedule.queue.task_index_queue);
-  kernel->schedule.queue.queue_position = num_active_tasks - 1;
+  kernel->schedule.queue.queue_position = (uint8_t)(num_active_tasks - 1);
 #if defined(VERBOSE_OUTPUT)
   print_queue(num_active_tasks);
 #endif  // defined(VERBOSE_OUTPUT)
